Refactor React News files
