--- src/imio/smartweb/core/webcomponents/src/components/News/ContactMap/ContactMap.scss
@@ -1,4 +1,0 @@
-.leaflet-container{
-    height: 100vh;
-}
-
